requirements.txt
linkify-it-py==2.1.0
markdown-it-py==4.0.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
numpy==2.4.4
pillow==12.2.0
platformdirs==4.9.4
pycryptodome==3.23.0
Pygments==2.20.0
rich==14.3.3
textual==8.2.1
typing_extensions==4.15.0
uc-micro-py==2.0.0
